+++ lib/spout/tests/json_helper.rb
@@ -1,0 +1,5 @@
+require 'json'
+
+def json_value(file, key)
+  begin JSON.parse(File.read(file))[key.to_s] rescue nil end
+end