+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.5.0
 
+### Enhancement
+- Tests added to check that variables of type choices have specified a domain
+
 ## 0.4.1 (August 16, 2013)
 
 ### Enhancement